start splitting code
